CHORE: Rename response_function.ts to response.ts
